Re Architecture of the Show Bottomsheet
