+++ lib/features/scanpay/store/scanpay_store.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'models/model_payusing.dart';
 import 'models/model_upi_data.dart';
 
+//TODO:  NO NEED OF STORE as of now
 class ScanpayStore extends ChangeNotifier {
   String amount = '';
   String message = '';
