+++ lib/features/transaction/store/transaction_store.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+//TODO:  NO NEED OF STORE as of now
 class TransactionStore extends ChangeNotifier {
   bool isExpense = true;
   String amount = '';
